integration-cli: move daemon stuff to its own file
